requirements.txt
biopython==1.84
requests
tqdm